+++ metnet/tests/test_fastcore.py
@@ -14,7 +14,7 @@ except ImportError:
 
 
 class TestFastcoreSimpleVlassisModel(unittest.TestCase):
-    """Test fastcore using the simple model in Vlassis et al. 2014."""
+    """Test fastcore using the simple model in [Vlassis14]_."""
 
     def setUp(self):
         # TODO use mock model instead of actual model
@@ -25,107 +25,130 @@ class TestFastcoreSimpleVlassisModel(unittest.TestCase):
         self.database.set_reaction('rxn_4', parse_reaction('|A| => |C|'))
         self.database.set_reaction('rxn_5', parse_reaction('|C| => |D|'))
         self.database.set_reaction('rxn_6', parse_reaction('|D| =>'))
-        self.model = MetabolicModel.load_model(self.database, self.database.reactions)
-        self.fastcore = fastcore.Fastcore(cplex.Solver())
+        self.model = MetabolicModel.load_model(
+            self.database, self.database.reactions)
+        self.solver = cplex.Solver()
 
     @unittest.skipIf(cplex is None, 'solver not available')
     def test_lp10(self):
-        result = self.fastcore.lp10(self.model, { 'rxn_6' }, { 'rxn_1', 'rxn_3', 'rxn_4', 'rxn_5' },
-                                    epsilon=0.001, scaling=1e3)
+        result = fastcore.lp10(self.model, { 'rxn_6' },
+                               { 'rxn_1', 'rxn_3', 'rxn_4', 'rxn_5' },
+                               solver=self.solver, epsilon=0.001, scaling=1e3)
         supp = set(fastcore.support(result, 0.999*0.001))
         self.assertEqual(supp, { 'rxn_1', 'rxn_3', 'rxn_6' })
 
     @unittest.skipIf(cplex is None, 'solver not available')
     def test_lp10_weighted(self):
         weights = { 'rxn_3': 1 }
-        result = self.fastcore.lp10(self.model, { 'rxn_6' }, { 'rxn_1', 'rxn_3', 'rxn_4', 'rxn_5' },
-                                    epsilon=0.001, scaling=1e3, weights=weights)
+        result = fastcore.lp10(self.model, { 'rxn_6' },
+                               { 'rxn_1', 'rxn_3', 'rxn_4', 'rxn_5' },
+                               solver=self.solver, epsilon=0.001, scaling=1e3,
+                               weights=weights)
         supp = set(fastcore.support(result, 0.999*0.001))
         self.assertEqual(supp, { 'rxn_1', 'rxn_3', 'rxn_6' })
 
         weights = { 'rxn_3': 3 }
-        result = self.fastcore.lp10(self.model, { 'rxn_6' }, { 'rxn_1', 'rxn_3', 'rxn_4', 'rxn_5' },
-                                    epsilon=0.001, scaling=1e3, weights=weights)
+        result = fastcore.lp10(self.model, { 'rxn_6' },
+                               { 'rxn_1', 'rxn_3', 'rxn_4', 'rxn_5' },
+                               solver=self.solver, epsilon=0.001, scaling=1e3,
+                               weights=weights)
         supp = set(fastcore.support(result, 0.999*0.001))
         self.assertEqual(supp, { 'rxn_1', 'rxn_4', 'rxn_5', 'rxn_6' })
 
     @unittest.skipIf(cplex is None, 'solver not available')
     def test_lp7(self):
-        result = self.fastcore.lp7(self.model, set(self.model.reactions), 0.001)
+        result = fastcore.lp7(self.model, set(self.model.reactions), 0.001,
+                              solver=self.solver)
         supp = set(fastcore.support_positive(result, 0.001*0.999))
         self.assertEqual(supp, { 'rxn_1', 'rxn_3', 'rxn_4', 'rxn_5', 'rxn_6' })
 
-        result = self.fastcore.lp7(self.model, {'rxn_5'}, 0.001)
+        result = fastcore.lp7(self.model, {'rxn_5'}, 0.001, solver=self.solver)
         supp = set(fastcore.support_positive(result, 0.001*0.999))
         self.assertEqual(supp, { 'rxn_1', 'rxn_4', 'rxn_5', 'rxn_6' })
 
     @unittest.skipIf(cplex is None, 'solver not available')
     def test_find_sparse_mode_singleton(self):
         core = { 'rxn_1' }
-        mode = set(self.fastcore.find_sparse_mode(self.model, core, set(self.model.reactions) - core,
-                                                    epsilon=0.001, scaling=1e3))
+        mode = set(fastcore.find_sparse_mode(
+            self.model, core, set(self.model.reactions) - core,
+            solver=self.solver, epsilon=0.001, scaling=1e3))
         self.assertEqual(mode, { 'rxn_1', 'rxn_3', 'rxn_6' })
 
         core = { 'rxn_2' }
-        mode = set(self.fastcore.find_sparse_mode(self.model, core, set(self.model.reactions) - core,
-                                                    epsilon=0.001, scaling=1e3))
+        mode = set(fastcore.find_sparse_mode(
+            self.model, core, set(self.model.reactions) - core,
+            solver=self.solver, epsilon=0.001, scaling=1e3))
         self.assertEqual(mode, set())
 
         core = { 'rxn_3' }
-        mode = set(self.fastcore.find_sparse_mode(self.model, core, set(self.model.reactions) - core,
-                                                    epsilon=0.001, scaling=1e3))
+        mode = set(fastcore.find_sparse_mode(
+            self.model, core, set(self.model.reactions) - core,
+            solver=self.solver, epsilon=0.001, scaling=1e3))
         self.assertEqual(mode, { 'rxn_1', 'rxn_3', 'rxn_6' })
 
         core = { 'rxn_4' }
-        mode = set(self.fastcore.find_sparse_mode(self.model, core, set(self.model.reactions) - core,
-                                                    epsilon=0.001, scaling=1e3))
+        mode = set(fastcore.find_sparse_mode(
+            self.model, core, set(self.model.reactions) - core,
+            solver=self.solver, epsilon=0.001, scaling=1e3))
         self.assertEqual(mode, { 'rxn_1', 'rxn_4', 'rxn_5', 'rxn_6' })
 
         core = { 'rxn_5' }
-        mode = set(self.fastcore.find_sparse_mode(self.model, core, set(self.model.reactions) - core,
-                                                    epsilon=0.001, scaling=1e3))
+        mode = set(fastcore.find_sparse_mode(
+            self.model, core, set(self.model.reactions) - core,
+            solver=self.solver, epsilon=0.001, scaling=1e3))
         self.assertEqual(mode, { 'rxn_1', 'rxn_4', 'rxn_5', 'rxn_6' })
 
         core = { 'rxn_6' }
-        mode = set(self.fastcore.find_sparse_mode(self.model, core, set(self.model.reactions) - core,
-                                                    epsilon=0.001, scaling=1e3))
+        mode = set(fastcore.find_sparse_mode(
+            self.model, core, set(self.model.reactions) - core,
+            solver=self.solver, epsilon=0.001, scaling=1e3))
         self.assertEqual(mode, { 'rxn_1', 'rxn_3', 'rxn_6' })
 
     @unittest.skipIf(cplex is None, 'solver not available')
     def test_find_sparse_mode_weighted(self):
         core = { 'rxn_1' }
         weights = { 'rxn_3': 1 }
-        mode = set(self.fastcore.find_sparse_mode(self.model, core, set(self.model.reactions) - core,
-                                                    epsilon=0.001, scaling=1e3, weights=weights))
+        mode = set(fastcore.find_sparse_mode(
+            self.model, core, set(self.model.reactions) - core,
+            solver=self.solver, epsilon=0.001, scaling=1e3, weights=weights))
         self.assertEqual(mode, { 'rxn_1', 'rxn_3', 'rxn_6' })
 
         weights = { 'rxn_3': 3 }
-        mode = set(self.fastcore.find_sparse_mode(self.model, core, set(self.model.reactions) - core,
-                                                    epsilon=0.001, scaling=1e3, weights=weights))
+        mode = set(fastcore.find_sparse_mode(
+            self.model, core, set(self.model.reactions) - core,
+            solver=self.solver, epsilon=0.001, scaling=1e3, weights=weights))
         self.assertEqual(mode, { 'rxn_1', 'rxn_4', 'rxn_5', 'rxn_6' })
 
     @unittest.skipIf(cplex is None, 'solver not available')
     def test_fastcc_inconsistent(self):
-        self.assertEqual(set(self.fastcore.fastcc(self.model, 0.001)), { 'rxn_2' })
+        self.assertEqual(
+            set(fastcore.fastcc(self.model, 0.001, solver=self.solver)),
+            { 'rxn_2' })
 
     @unittest.skipIf(cplex is None, 'solver not available')
     def test_fastcc_is_consistent_on_inconsistent(self):
-        self.assertFalse(self.fastcore.fastcc_is_consistent(self.model, 0.001))
+        self.assertFalse(fastcore.fastcc_is_consistent(
+            self.model, 0.001, solver=self.solver))
 
     @unittest.skipIf(cplex is None, 'solver not available')
     def test_fastcc_is_consistent_on_consistent(self):
         self.model.remove_reaction('rxn_2')
-        self.assertTrue(self.fastcore.fastcc_is_consistent(self.model, 0.001))
+        self.assertTrue(fastcore.fastcc_is_consistent(
+            self.model, 0.001, solver=self.solver))
 
     @unittest.skipIf(cplex is None, 'solver not available')
     def test_fastcc_consistent_subset(self):
-        self.assertEqual(self.fastcore.fastcc_consistent_subset(self.model, 0.001), set(['rxn_1', 'rxn_3', 'rxn_4', 'rxn_5', 'rxn_6']))
+        self.assertEqual(fastcore.fastcc_consistent_subset(
+            self.model, 0.001, solver=self.solver),
+            set(['rxn_1', 'rxn_3', 'rxn_4', 'rxn_5', 'rxn_6']))
 
     @unittest.skipIf(cplex is None, 'solver not available')
     def test_fastcore_global_inconsistent(self):
         self.database.set_reaction('rxn_7', parse_reaction('|E| <=>'))
-        with self.assertRaises(Exception):
-            self.fastcore.fastcore(self.model, { 'rxn_7' }, 0.001)
+        self.model.add_reaction('rxn_7')
+        with self.assertRaises(fastcore.FastcoreError):
+            fastcore.fastcore(self.model, { 'rxn_7' }, 0.001,
+                              solver=self.solver)
 
 
 class TestFastcoreTinyBiomassModel(unittest.TestCase):
@@ -135,23 +158,29 @@ class TestFastcoreTinyBiomassModel(unittest.TestCase):
         # TODO use mock model instead of actual model
         self.database = DictDatabase()
         self.database.set_reaction('rxn_1', parse_reaction('=> |A|'))
-        self.database.set_reaction('rxn_2', parse_reaction('(0.000001) |A| =>'))
-        self.model = MetabolicModel.load_model(self.database, self.database.reactions)
-        self.fastcore = fastcore.Fastcore(cplex.Solver())
+        self.database.set_reaction('rxn_2',
+            parse_reaction('(0.000001) |A| =>'))
+        self.model = MetabolicModel.load_model(
+            self.database, self.database.reactions)
+        self.solver = cplex.Solver()
 
     @unittest.skipIf(cplex is None, 'solver not available')
     def test_fastcc_is_consistent(self):
-        self.assertTrue(self.fastcore.fastcc_is_consistent(self.model, 0.001))
+        self.assertTrue(fastcore.fastcc_is_consistent(
+            self.model, 0.001, solver=self.solver))
 
     @unittest.skipIf(cplex is None, 'solver not available')
     def test_fastcore_induced_model(self):
         core = { 'rxn_2' }
-        self.assertEquals(set(self.fastcore.fastcore(self.model, core, 0.001)), { 'rxn_1', 'rxn_2' })
+        self.assertEquals(set(fastcore.fastcore(
+            self.model, core, 0.001, solver=self.solver)),
+            { 'rxn_1', 'rxn_2' })
 
     @unittest.skipIf(cplex is None, 'solver not available')
     def test_fastcore_induced_model_high_epsilon(self):
         core = { 'rxn_2' }
-        self.assertEquals(set(self.fastcore.fastcore(self.model, core, 0.1)), { 'rxn_1', 'rxn_2' })
+        self.assertEquals(set(fastcore.fastcore(
+            self.model, core, 0.1, solver=self.solver)), { 'rxn_1', 'rxn_2' })
 
 
 class TestFlippingModel(unittest.TestCase):
@@ -164,14 +193,16 @@ class TestFlippingModel(unittest.TestCase):
         self.database.set_reaction('rxn_2', parse_reaction('|A| <=> |B|'))
         self.database.set_reaction('rxn_3', parse_reaction('|C| <=> |B|'))
         self.database.set_reaction('rxn_4', parse_reaction('|C| <=>'))
-        self.model = MetabolicModel.load_model(self.database, self.database.reactions)
-        self.fastcore = fastcore.Fastcore(cplex.Solver())
+        self.model = MetabolicModel.load_model(
+            self.database, self.database.reactions)
+        self.solver = cplex.Solver()
 
     @unittest.skipIf(cplex is None, 'solver not available')
     def test_fastcore_induced_model(self):
         core = { 'rxn_2', 'rxn_3' }
-        self.assertEquals(set(self.fastcore.fastcore(self.model, core, 0.001)),
-                            { 'rxn_1', 'rxn_2', 'rxn_3', 'rxn_4' })
+        self.assertEquals(set(
+            fastcore.fastcore(self.model, core, 0.001, solver=self.solver)),
+            { 'rxn_1', 'rxn_2', 'rxn_3', 'rxn_4' })
 
 
 if __name__ == '__main__':